restructure code
