--- src/com/mocean/exception/MoceanErrorException.java
@@ -1,20 +1,0 @@
-package com.mocean.exception;
-
-import com.mocean.modules.mapper.ErrorResponse;
-
-public class MoceanErrorException extends Exception {
-    private ErrorResponse errorResponse;
-
-    public MoceanErrorException(String errMsg) {
-        super(errMsg);
-    }
-
-    public MoceanErrorException(ErrorResponse errorResponse) {
-        super(errorResponse.getRawResponse());
-        this.errorResponse = errorResponse;
-    }
-
-    public ErrorResponse getErrorResponse() {
-        return errorResponse;
-    }
-}
