+++ src/main/java/com/mocean/modules/ErrorResponse.java
@@ -1,0 +1,21 @@
+package com.mocean.modules;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import javax.xml.bind.annotation.XmlElement;
+
+public class ErrorResponse extends AbstractResponse {
+    @XmlElement(name = "err_msg")
+    @JsonProperty("err_msg")
+    private String errMsg;
+
+    public String getErrMsg() {
+        return errMsg;
+    }
+
+    @Override
+    public ErrorResponse setRawResponse(String rawResponse) {
+        super.rawResponse = rawResponse;
+        return this;
+    }
+}
