+++ src/main/java/com/mocean/utils/Utils.java
@@ -1,0 +1,11 @@
+package com.mocean.utils;
+
+public class Utils {
+    public static boolean isNullOrEmpty(String str) {
+        return str == null || str.trim().isEmpty();
+    }
+
+    public static boolean isArray(Object obj) {
+        return obj != null && obj.getClass().isArray();
+    }
+}
